+++ src/index.js
@@ -40,3 +40,4 @@ export { default as FileField } from "./components/form/FileField";
 export { default as ImageField } from "./components/form/ImageField";
 export { default as SelectField } from "./components/form/SelectField";
 export { default as SubmitButton } from "./components/form/SubmitButton";
+export { default as TextField } from "./components/form/TextField";