+++ test/ImageEditor.test.js
@@ -90,41 +90,6 @@ test("can click save to save", async () => {
   expect(response).not.toBe(null);
 });
 
-test("can hit enter to save", async () => {
-  let response = null;
-  const onEdit = edited => {
-    response = edited;
-  };
-
-  const component = await mountEditor(onEdit);
-
-  component.instance().cropper.getCroppedCanvas = () => ({
-    toDataURL() {
-      return image;
-    }
-  });
-
-  component.instance().handleKeyPressed({
-    key: "Enter",
-    preventDefault() {},
-    stopPropagation() {}
-  });
-
-  expect(response).not.toBe(null);
-});
-
-test("does not react to other keys", async () => {
-  let response = null;
-  const onEdit = edited => {
-    response = edited;
-  };
-
-  const component = await mountEditor(onEdit);
-  component.instance().handleKeyPressed({ key: "Shift" });
-
-  expect(response).toBe(null);
-});
-
 test("does not attempt to set state if already unmounted", async () => {
   const component = await mountEditor();